requirements.txt
flask
PyPDF2
pymongo
flask_cors
google.generativeai
langchain_google_genai
dotenv